Moved authentication functionality to LoginUtils.js
